Handle subver incompatibilities
Due to different length subver between the old and new versions of Bitcoin Plus bespoke subver handling is implemented here. Blank subver is sent to old clients to prevent an unhandled exception and in this version subver is based on PROTOCOL_VERSION. 80003 - 80008 reserved for 2.6.0 - 2.6.5. 80010 - 80015 reserved for 2.7.0 - 2.7.5. 80016 - 80021 reserved for 2.8.0 - 2.8.5. Anything over 80022 is assumed to be sending subver as normal.